+++ src/test/ValentinaTest/tst_vtextstream.cpp
@@ -1289,7 +1289,7 @@ void TST_VTextStream::pos3LargeFile()
         // NOTE: The unusual spacing is to ensure non-1-character whitespace.
         QString lineString = QStringLiteral(" 0  1  2\t3  4\t \t5  6  7  8   9 \n");
         // Approximate 50kb text file
-        const int NbLines = (50 * 1024) / lineString.length() + 1;
+        const int NbLines = (50 * 1024) / static_cast<int>(lineString.length()) + 1;
         for (int line = 0; line < NbLines; ++line)
             out << lineString;
         // File is automatically flushed and closed on destruction.
